+++ podonos/__init__.py
@@ -2,6 +2,6 @@ from .core.file import File
 from .entity.flash_eval import FlashEvalResult
 from .sdk import Podonos, Client
 
-__version__ = "0.33.2"
+__version__ = "0.33.3"
 
 init = Podonos.init